+++ src/main/java/org/snakegame/v1/board/Board.java
@@ -1,7 +1,7 @@
-package org.snakegame.board;
+package org.snakegame.v1.board;
 
-import org.snakegame.cell.Cell;
-import org.snakegame.constants.CellState;
+import org.snakegame.v1.cell.Cell;
+import org.snakegame.v1.constants.CellState;
 
 import java.util.ArrayList;
