+++ src/main/java/org/snakegame/v1/expception/BoardNotEmpty.java
@@ -1,4 +1,4 @@
-package org.snakegame.expception;
+package org.snakegame.v1.expception;
 
 public class BoardNotEmpty extends RuntimeException {
     public BoardNotEmpty(String message) {
